refactor: separte main package into server package and main package.
